@@ -59,13 +59,11 @@
     "postcss-flexbugs-fixes": "^3.0.0",
     "postcss-loader": "^2.0.6",
     "postcss-scss": "^1.0.2",
-    "postgraphql": "^3.3.0",
-    "preact": "^8.2.1",
-    "preact-compat": "^3.16.0",
+    "postgraphql": "4.0.0-alpha2.0",
     "prop-types": "^15.5.10",
-    "react": "^15.0.0",
+    "react": "^16.0.0",
     "react-apollo": "^1.4.2",
-    "react-dom": "^15.0.0",
+    "react-dom": "16.0.0-beta.1",
     "react-hot-loader": "next",
     "react-redux": "^5.0.5",
     "react-router": "^4.1.1",
